feat: http error handling | relocate HttpErrorPayload to a better place
NO-TICKET
